Move compression classes to separate file
Handle fstream construction/destruction in streambuf classes.
This also avoids the need for storing pointers in the pword() array.
