+++ core/include/core/G3Writer.h
@@ -9,7 +9,8 @@
 class G3Writer : public G3Module {
 public:
 	G3Writer(std::string filename,
-	    std::vector<G3Frame::FrameType> streams={}, bool append=false);
+	    std::vector<G3Frame::FrameType> streams={}, bool append=false,
+	    size_t buffersize=1024*1024);
 	// Writes to file <filename> all frames with types in <streams>.
 	// If <streams> is empty (default), writes all frames.
 	virtual ~G3Writer();
