+++ core/src/G3Writer.cxx
@@ -4,11 +4,11 @@
 
 G3Writer::G3Writer(std::string filename,
     std::vector<G3Frame::FrameType> streams,
-    bool append) :
+    bool append, size_t buffersize) :
     filename_(filename), stream_(nullptr), streams_(streams)
 {
 	g3_check_output_path(filename);
-	g3_ostream_to_path(stream_, filename, append);
+	g3_ostream_to_path(stream_, filename, append, buffersize);
 }
 
 G3Writer::~G3Writer()
@@ -39,7 +39,7 @@ void G3Writer::Process(G3FramePtr frame, std::deque<G3FramePtr> &out)
 void G3Writer::Flush()
 {
 	try {
-		g3_ostream_flush(stream_);
+		stream_.flush();
 	} catch (...) {
 	}
 }
@@ -56,9 +56,10 @@ PYBINDINGS("core") {
 	      "types to the second optional argument (streams). If no streams argument "
 	      "is given, writes all types of frames. If append is set to True, will "
 	      "append frames to its output file rather than overwriting it.",
-        init<std::string, std::vector<G3Frame::FrameType>, bool>((arg("filename"),
-	    arg("streams")=std::vector<G3Frame::FrameType>(), arg("append")=false)))
-        .def("Flush", &G3Writer::Flush)
-        .def_readonly("__g3module__", true)
+	init<std::string, std::vector<G3Frame::FrameType>, bool, size_t>((arg("filename"),
+	    arg("streams")=std::vector<G3Frame::FrameType>(), arg("append")=false,
+	    arg("buffersize")=1024*1024)))
+	.def("Flush", &G3Writer::Flush)
+	.def_readonly("__g3module__", true)
 	;
 }
